+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `Innmind\Server\Control\Server\Process::output()` now returns `Innmind\Immutable\Sequence<Innmind\Server\Control\Server\Process\Output\Chunk>`
+- `Innmind\Server\Control\Server\Process` is now a final class, old implementations are now declared internal
 
 ### Removed